+++ src/aiologic/_queue.py
@@ -21,8 +21,7 @@ from .lowlevel import (
     green_checkpoint,
     lazydeque,
 )
-from .lowlevel._utils import _copies as copies
-from .meta import MISSING, MissingType
+from .meta import MISSING, MissingType, copies
 
 if sys.version_info >= (3, 11):
     from typing import overload