+++ src/aiologic/_testing/_exceptions.py
@@ -7,8 +7,7 @@ from __future__ import annotations
 
 import sys
 
-from aiologic.lowlevel._utils import _replaces as replaces
-from aiologic.meta import DEFAULT, DefaultType
+from aiologic.meta import DEFAULT, DefaultType, replaces
 
 from ._executors import TaskExecutor, current_executor