+++ src/aiologic/lowlevel/_safety.py
@@ -13,10 +13,9 @@ from typing import TYPE_CHECKING, Any, Literal, TypeVar
 
 from wrapt import ObjectProxy, decorator
 
-from aiologic.meta import MISSING, MissingType
+from aiologic.meta import MISSING, MissingType, replaces
 
 from ._threads import current_thread_ident
-from ._utils import _replaces as replaces
 
 if sys.version_info >= (3, 11):
     from typing import overload